requirements.txt
# RestaurantOS — Dependencies
# Instalar con: pip install -r requirements.txt

# Core
streamlit>=1.35.0
python-dotenv>=1.0.0

# Supabase
supabase>=2.4.0
postgrest>=0.16.0

# Google Gemini (OCR de facturas) — nuevo SDK oficial
google-genai>=1.0.0

# HTTP client (Loyverse API)
requests>=2.31.0

# Image processing
Pillow>=10.3.0
PyMuPDF>=1.24.0

# Data manipulation
pandas>=2.2.0

# (Opcional) Para análisis avanzado y gráficos
plotly>=5.20.0
